requirements.txt
tweepy==3.5.0
influxdb==5.0.0
pyyaml==3.12.0
